Implement more robust default config copying Move JsonFileDict to config_manager Fix commands not working lol
